+++ backend/Routers/Curation.js
@@ -1,23 +1,24 @@
 import express from 'express';
 import { curationNestedCommentRouter } from './Comment.js';
 // prettier-ignore
-import { getCurationList, postCuration, putCuration, deleteCuration } from '../Services/CurationService.js';
-// prettier-ignore
 import { validatePostCuration, validateUpdateCuration, validateDeleteCuration } from '../Validators/CurationValidator.js';
 import hashingPassword from '../Middlewares/hashing.js';
+import asyncHandler from '../Middlewares/asyncHandler.js';
+import CurationController from '../Controllers/CurationController.js';
+
 const styleNestedCurationRouter = express.Router({ mergeParams: true });
 const CurationRouter = express.Router();
 
 // prettier-ignore
 styleNestedCurationRouter
   .route('/')
-  .post(validatePostCuration, hashingPassword(), postCuration())
-  .get(getCurationList());
+  .post(validatePostCuration, hashingPassword(), asyncHandler(CurationController.postCuration))
+  .get(asyncHandler(CurationController.getCurationList));
 
 // prettier-ignore
 CurationRouter.route('/:id')
-  .put(validateUpdateCuration, putCuration())
-  .delete(validateDeleteCuration, deleteCuration());
+  .put(validateUpdateCuration, asyncHandler(CurationController.putCuration))
+  .delete(validateDeleteCuration, asyncHandler(CurationController.deleteCuration));
 
 CurationRouter.use('/:curationId/comments', curationNestedCommentRouter);
